added popup icon

+++ generate-fonts.sh
@@ -1,9 +1,14 @@
 #sessionId=$(curl --form "config=@fontello-conf.json" http://fontello.com/)
-echo $sessionId
+#echo $sessionId
 #wget http://fontello.com/$sessionId/get -O fonts.zip
-unzip fonts.zip -d ./fontello
-dir=$(ls ./fontello)
-cp -r ./fontello/$dir/font ./static
-cp -r ./fontello/$dir/css ./static
-cp ./fontello/$dir/demo.html ./static
-rm -r ./fontello/
+if [  -f "$1" ]; then
+    unzip $1 -d ./fontello
+    dir=$(ls ./fontello)
+    cp -r ./fontello/$dir/font ./static
+    cp -r ./fontello/$dir/css ./static
+    cp ./fontello/$dir/demo.html ./static
+    cp ./fontello/$dir/config.json ./
+    rm -r ./fontello/
+else
+ echo "Pass path to archive downloaded from http://fontello.com/ as first argument"
+fi
